abstracted codepen functionality into codepen component, simplified examplecard
