+++ packages/example-app-base/src/components/CodepenComponent/CodepenComponent.tsx
@@ -1,0 +1,51 @@
+import * as React from 'react';
+import { css } from 'office-ui-fabric-react/lib/Utilities';
+import '../ExampleCard/ExampleCard.scss';
+import { CommandButton } from 'office-ui-fabric-react/lib/Button';
+
+export interface ICodepenProps {
+  /* JS string to be passed into Codepen */
+  jsContent: string;
+}
+
+export class CodepenComponent extends React.Component<ICodepenProps> {
+  constructor(props: ICodepenProps) {
+    super(props);
+  }
+
+  public render(): JSX.Element {
+    const { jsContent } = this.props;
+
+    // boilerplate for codepen API
+    const htmlContent =
+      '<script src="//unpkg.com/office-ui-fabric-react/dist/office-ui-fabric-react.min.js"></script>\n<div id=\'content\'></div>';
+
+    const headContent =
+      '<script type="text/javascript" src="https://unpkg.com/react@16/umd/react.development.js"></script>\n<script type="text/javascript" src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"></script>';
+
+    const valueData = {
+      title: 'Fabric Example Pen',
+      html: htmlContent,
+      head: headContent,
+      js: jsContent,
+      js_pre_processor: 'typescript'
+    };
+
+    // reformat the JSON string to take out the quotes so it'll work with the Codepen API
+    const JSONstring = JSON.stringify(valueData)
+      .replace(/"/g, '&quot;')
+      .replace(/'/g, '&apos;');
+
+    return (
+      <form action="https://codepen.io/pen/define" method="POST" target="_blank">
+        <input type="hidden" name="data" value={JSONstring} />
+        <CommandButton
+          type="submit"
+          iconProps={{ iconName: 'OpenInNewWindow' }}
+          text="Export to Codepen"
+          className={css('ExampleCard-codeButton')}
+        />
+      </form>
+    );
+  }
+}
