+++ packages/example-app-base/src/components/ExampleCard/ExampleCard.tsx
@@ -3,13 +3,14 @@ import { css } from 'office-ui-fabric-react/lib/Utilities';
 import './ExampleCard.scss';
 import { CommandButton } from 'office-ui-fabric-react/lib/Button';
 import { Highlight } from '../Highlight/Highlight';
+import { CodepenComponent } from '../CodepenComponent/CodepenComponent';
 
 export interface IExampleCardProps {
   /* Example Title */
   title: string;
   /* Experimental Component? */
   isOptIn?: boolean;
-  /* Example Code */
+  /* Example Code as a string */
   code?: string;
   /* Children of the Example */
   children?: React.ReactNode;
@@ -22,18 +23,12 @@ export interface IExampleCardProps {
   /* Example is scrollable ? */
   isScrollable?: boolean;
   /* JS string for Codepen portion of Example */
-  codepen?: string;
+  codepenJS?: string;
 }
 
 export interface IExampleCardState {
   isCodeVisible?: boolean;
 }
-
-// boilerplate for codepen API
-const htmlContent = `<script src="//unpkg.com/office-ui-fabric-react/dist/office-ui-fabric-react.min.js"></script>
-<div id=\'content\'></div>`;
-
-const headContent = `<script type="text/javascript" src="https://unpkg.com/react@16/umd/react.development.js"></script><script type="text/javascript" src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"></script>`;
 
 export class ExampleCard extends React.Component<IExampleCardProps, IExampleCardState> {
   constructor(props: IExampleCardProps) {
@@ -47,21 +42,7 @@ export class ExampleCard extends React.Component<IExampleCardProps, IExampleCard
   }
 
   public render(): JSX.Element {
-    const { title, code, children, isRightAligned = false, isScrollable = true, codepen } = this.props;
-
-    // more codepen API boilerplate
-    const jsContent = codepen;
-    const valueData = {
-      title: 'Fabric Example Pen',
-      html: htmlContent,
-      head: headContent,
-      js: jsContent,
-      js_pre_processor: 'typescript'
-    };
-    // reformat the JSON string to take out the quotes so it'll work with the Codepen API
-    const JSONstring = JSON.stringify(valueData)
-      .replace(/"/g, '&quot;')
-      .replace(/'/g, '&apos;');
+    const { title, code, children, isRightAligned = false, isScrollable = true, codepenJS } = this.props;
 
     const { isCodeVisible } = this.state;
     let rootClass = 'ExampleCard' + (this.state.isCodeVisible ? ' is-codeVisible' : '');
@@ -71,19 +52,7 @@ export class ExampleCard extends React.Component<IExampleCardProps, IExampleCard
         <div className="ExampleCard-header">
           <span className="ExampleCard-title ms-font-l">{title}</span>
           <div className="ExampleCard-toggleButtons ms-font-l">
-            {codepen ? (
-              <form action="https://codepen.io/pen/define" method="POST" target="_blank">
-                <input type="hidden" name="data" value={JSONstring} />
-                <CommandButton
-                  type="submit"
-                  iconProps={{ iconName: 'OpenInNewWindow' }}
-                  text="Export to Codepen"
-                  className={css('ExampleCard-codeButton')}
-                />
-              </form>
-            ) : (
-              undefined
-            )}
+            {codepenJS ? <CodepenComponent jsContent={codepenJS} /> : undefined}
             {code ? (
               <CommandButton
                 iconProps={{ iconName: 'Embed' }}
